Fix build on macOS Catalina
The Secure Transport framework is deprecated in the macOS Catalina SDK
and generates '-Wdeprecated-declarations' warnings which cause the
build to fail because of '-Werror'. This commit fixes the build by
setting the minimum target and SDK to macOS 10.13. Switching to the
OpenSSL backend seems to be the only feasible long-term solution,
since the new Network framework provides only a higher-level API.